Testing
git-svn-id: https://ssl.drqueue.org/svn/branches/0.64.x@1611 94d6b6b3-0ee6-0310-ab95-9b4161196ee2

+++ libdrqueue/semaphores.c
@@ -73,7 +73,7 @@ semaphore_release (int64_t semid) {
 
 int
 semaphore_valid (int64_t semid) {
-  if (semctl ((int)semid,0,GETVAL) == -1) {
+  if (semctl ((int64_t)semid,0,GETVAL) == -1) {
     drerrno_system = errno;
     log_auto (L_DEBUG,"semaphore_valid(): not valid '%ji'. (Msg: %s)",semid,strerror(drerrno_system));
     return 0;
